Delete LICENSE
Date_Query and Compare are currently derivatives of work from WordPress itself, meaning the GPLv2.0 or higher license must remain.

This MIT license is being deleted, to be replaced with the correct GPL license.

If such a time arrises where Date_Query and Compare can be wholly replaced, the MIT license should be reconsidered.