Set db_global to 'wpdb' by default.
This avoids more complicated dependency injection being a requirement, simply to setup the very basic database connection that is required to make anything work.

A future version of BerlinDB\Base may abstract this even further, but because WordPress is the only natively supported environment currently, there is no reason to make this anymore difficult than it needs to be today.