Introduce set_defaults() method.
This commit introduces a private data() array, and a public method for classes that extend Base to call to setup the defaults.

This is necessary to make sure that all default class variables make their way into the data array so that magic methods can access them like they can any other variable.