Updates delete_metadata_by_mid to use table name, not the parsed meta table name. #51
This change resolves this problem because `delete_metadata_by_mid` uses
`_get_meta_table` to retrieve the meta table from the database. This
function automatically appends `meta` to the table name, and does not
need a prefix.

Due to this, there is no need to do anything but directly pass
`table_name` to `delete_metadata_by_mid`