Updates upgrade filter to use the upgrade version instead of the callback (#54)
* Updates upgrade filter to use the upgrade version instead of the callback
* Refactors upgrade filter method to use a foreach loop. #53
* Corrects variable call on upgrades. #53
* Extracts get upgrades loop into its own method. #54

This is because BerlinDB is ARRAY_FILTER_USE_KEY requires PHP 5.6 or higher. Many plugins that rely on BerlinDB have a minimum requirement of 5.2. Due to this, the best option was to rework this filter to use a foreach loop.

This makes the code easier to read, and will help ensure the code is readily test-able should unit tests be implemented in the future.