Nulls: validate null values if allowed, on any column.
Up until now, the "Column::allow_null" flag was only partially implemented, working sometimes for "datetime" columns, but not for any others.

This commit ensures that, in the "Query::validate_item" method, null values for all item keys are allowed when "allow_null" is true for that column, and returns false early when attempting to save a null value to an item key when it is not allowed.

This commit also removes the special datetime, column type, null value handling that was added as part of #22, since it is now explicitly globally allowed (when allowed) in the global validator.

This commit also modifies "Column::sanitize_default" to only allow null defaults when "allow_null" is true, adds explicit support for strings and integers, and adds a @todo regarding other column types that may have their own unique default values (datetime, decimal, timestamp, and others come to mind).

Fixes #76.