Date: merge $now abstraction from Sugar Calendar.
This change abstracts the `$now = time()` portion of build_mysql_datetime() out into a method parameter, allowing for $now to be passed in if needed for consistency between query clauses.

It also adds support for string-or-int, non-array $datetime values, as well as the "Y-m-d H:i:s" format directly if identified.