Query: only update relevant columns in update_item().
This change ensures that only the columns being passed updated data are actually updated, avoiding updating columns where no update data was passed in.

The exception to this rule continues to be any `modified` column type, which is automatically updated on calls to update_item().

Along with this change come a few optimiziations:

* Bail early if no data is passed in to update the row
* Bail at midway if no save or meta data exists after compared with relevant columns
* Only update meta-data after successful update. This continues to allow for meta-only updates by way of a successful `modified` column update. If no `modified` column exists, this is a behavior change and will need to be investigated.