Release/2.0.0 (#57)
Merging this a little early so we can get Composer support in a stable version. I've been running this for a few months now and it seems to work fine.

* Move classes to PSR-4 compliant src structure

* Add composer, autoloader support & gitignore

* Add use to fix missing class issue after file move

* Date: Remove duplicate "use" statement.

* Query: non-functional clean-up.

This change normalizes a bunch of variable names, docs, punctuation, and aliased functions (like join vs implode) - consistency!

* Query: add copy_item() method. (#89)

This change introduces a way to allow for copying an existing item/Row in the database, with the option to override that data via the $data parameter.

The goal of including this method as a low-level API is to reduce the amount of code necessary to support this action higher up in the application layer (global functions, etc...)

* Table: fix format of query inside index_exists(). (#90)

This change prevents quotes around the $column value by not preparing it, and also by making sure its value is only 1 of 2 intended columns: Key_name and Column_name.

This fixes incorrectly generated SQL, allowing this method to function as intended instead of incorrectly returning no results.

Props @ashleyfae. Fixes #87.

* First pass autoloader for non-Composer setups.

Props szepeviktor.

* Check if 'where' is empty before using #101 (#102)

* Check if 'where' is empty before using #101
* Added protection against empty "where" clauses #101

Co-authored-by: John James Jacoby <johnjamesjacoby@me.com>

Co-authored-by: danieliser <danieliser@wizardinternetsolutions.com>
Co-authored-by: Ashley Gibson <ashley@nosegraze.com>