Query: always use `set_last_changed()` in `update_last_changed_cache()`
This change ensures that anytime an item is updated or deleted, `last_changed` will also be set to the current microtime.

This fixes a bug where re-querying for items with the same `Query` instance that previously updated or deleted an item would return results that were out-of-date with the database, because `last_changed` was only ever set a single time.

Props spencerfinnell.

Fixes #160.