Query: reset `fields` when counting. (#177)
This change prevents unintended behavior when both the `count` and `fields` query variables are non-defaults, causing the SQL that is generated to be incomplete.

By resetting `fields` in `parse_query()`, we prioritize doing the standard full `COUNT(*)`.

(It is possible in the future that this gets reverted and both `count` and `fields` are supported, allowing for getting a count of any specific combination of columns/fields.)

Props robincornett.

Fixes #168.